requirements.txt
langchain
streamlit
ibm_watson_machine_learning
pypdf
sentence-transformers
chromadb
watchdog
langchain-ibm
langchain-community
python-docx 
python-pptx